+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - update k8s.io packages to v0.28.11
+- introduced a caching mechanism to avoid multiple calls to the remote api-server
 
 ### Added